mcp-sse-gdrive/pyproject.toml
[project]
name = "mcp-sse-gdrive"
version = "0.1.0"
description = "Google Drive SSE MCP Server"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-multipart>=0.0.5",
    "google-auth-oauthlib>=0.4.6",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
    "fastmcp>=0.1.0",
    "httpx>=0.24.0"
]

[project.scripts]
mcp-sse-gdrive = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_sse_gdrive"]

[tool.hatch.metadata]
allow-direct-references = true